requirements.txt
flask==3.0.0
flask-restful==0.3.10
flask-cors==4.0.0
boto3==1.34.60
python-dotenv==1.0.1
gunicorn==22.0.0
apispec==6.3.0
marshmallow==3.20.1
pymysql==1.1.0
sqlalchemy==2.0.28
pandas==2.1.4
cryptography==41.0.7
tabulate==0.9.0